refactor: compose tool registry, unify mcp-remote invocation
The registry no longer extends Map, so the inherited mutators cannot
desync its bookkeeping. The npx mcp-remote invocation lives in one
builder shared by the config writer, copy button, and .mcpb manifest.
Drop the unused cosineSimilarity export and stop re-checking the log
directory on every dev log line.
