+++ packages/obsidian-plugin/src/features/mcp-client-config/services/mcpbGenerator.ts
@@ -1,5 +1,6 @@
 import { zipSync, strToU8 } from "fflate";
 import { ICON_PNG_B64 } from "../assets/iconPng";
+import { mcpRemoteInvocation } from "./generators";
 
 export type McpbGeneratorInput = {
   version: string;
@@ -36,16 +37,7 @@ function buildManifest(input: McpbGeneratorInput): Record<string, unknown> {
     server: {
       type: "node",
       entry_point: "server/index.js",
-      mcp_config: {
-        command: "npx",
-        args: [
-          "-y",
-          "mcp-remote",
-          url,
-          "--header",
-          `Authorization: Bearer ${input.token}`,
-        ],
-      },
+      mcp_config: mcpRemoteInvocation(url, input.token),
     },
     // No user_config: token and port are embedded so Claude Desktop
     // installs with zero prompts. This bundle is generated by the plugin
